Validating report headers and checking if columns are not empty
