+++ features/reports.feature
@@ -141,9 +141,40 @@ Scenario Outline: User can download the report
   Then the report is downloaded successfully
 
   Examples:
-  |vaccineType                        | site          | day                             |
-  | COVID-19                          | Albert House  | Last 31 days (includes today)   |
-  | Flu                               | Albert House  | Last 31 days (includes today)   |
-  | Pertussis                         | Albert House  | Last 7 days (includes today)    |
-  | Respiratory syncytial virus (RSV) | Albert House  | Last 14 days (includes today)   |
-  | COVID-19                          | Albert House  | Last 31 days (includes today)   |
+  |vaccineType                        | site          | day                                       |
+  | COVID-19                          | Albert House  | Last 31 days (includes today)             |
+  | Flu                               | Albert House  | Last 31 days (includes today)             |
+  | Pertussis                         | Albert House  | Last 7 days (includes today)              |
+  | Respiratory syncytial virus (RSV) | Albert House  | Last 14 days (includes today)             |
+  | COVID-19                          | Albert House  | Last 31 days (includes today)             |
+  | COVID-19                          | Albert House  | Select a custom date range up to 31 days  |
+
+  Scenario Outline: Record a vaccine and generate a report for today
+    Given I am logged into the RAVS app
+    And I login to RAVS and set vaccinator details with <site> and <care_model> and get patient details for <nhs_number> with option <index> and choose to vaccinate with vaccine details as <chosen_vaccine>, <batch_number> with <batch_expiry_date>
+    And I search for a patient with the NHS number in the find a patient screen
+    And I open the patient record by clicking on patient <name>
+    When I click choose vaccine button and choose the <chosen_vaccine>, <batch_number> with <batch_expiry_date> and click continue
+    And I assess the patient's <eligibility> with the details and date as <assess_date> and click continue to record consent screen button
+    And I record <consent> with the details and click continue to vaccinate button
+    And I record <vaccination> details and date as <vaccination_date> and click Continue to Check and confirm screen
+    Then I need to be able to see the patient <name>, <dob>, <address> and vaccination details on the check and confirm screen
+    And when I click confirm and save button, I should see a record saved dialogue
+    And I search for a patient with the NHS number in the find a patient screen
+    And I open the patient record by clicking on patient <name>
+    And the immunisation history of the patient should be updated in the patient details page and not be deleted
+    When I click the reports navigation link
+    And I click the create report button
+    And I click the Today radio button and click Continue
+    And I select the vaccine type <chosen_vaccine> and click continue
+    And I select the site <site> and click continue
+    And I click continue on the data page
+    And I click Confirm and create report button in the check and confirm page
+    And I click download report button
+    Then the report is downloaded successfully and contains the vaccine record for <nhs_number>
+
+  Examples:
+    | index | nhs_number | site | care_model | eligibility | assess_date | consent | vaccination | vaccination_date | name    | dob        | address                                       | chosen_vaccine | batch_number     | batch_expiry_date |
+    | 0  | 9693632109 | Albert House | Vaccination Centre open to the public | yes        | today      | yes     | yes         | today | Bill GARTON | 23/6/1946 | 1 MOUNT AVENUE, BARTON-UPON-HUMBER, S HUMBERSIDE, DN18 5DW | COVID-19  | AUTOMATION-SJ1   | 19/10/2026   |
+    | 0  | 9693632109 | Albert House | Vaccination Centre open to the public | yes        | today      | yes     | yes         | today-32 | Bill GARTON | 23/6/1946 | 1 MOUNT AVENUE, BARTON-UPON-HUMBER, S HUMBERSIDE, DN18 5DW | COVID-19  | AUTOMATION-SJ1   | 19/10/2026   |
+
